commit script updates
