+++ packages/graphql/package.json
@@ -76,6 +76,7 @@
     "ts-node": "10.9.1",
     "ts-node-dev": "2.0.0",
     "tsup": "7.2.0",
+    "tsx": "4.9.1",
     "typescript": "5.1.6",
     "web-push": "3.6.1"
   },
